# public/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Board API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/Profile"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": [],
                        "OAuth2": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ProfileEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "Bearer": [],
                        "OAuth2": []
                    }
                ]
            }
        }
    },
    "definitions": {
...
        "ProfileEditRequest": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "Profile": {
            "properties": {
                "id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "phone": {
                    "properties": {
                        "number": {
                            "type": "string"
                        },
                        "verified": {
                            "type": "boolean"
                        }
                    },
                    "type": "object"
                },
                "name": {
                    "properties": {
                        "first": {
                            "type": "string"
                        },
                        "last": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                }
            },
            "type": "object"
        }
    },
    "securityDefinitions": {
        "OAuth2": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://localhost:8080/oauth/token"
        },
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}